pyproject.toml
[tool.poetry]
name = "goit-web-homework-11"
version = "0.1.0"
description = ""
authors = ["Yuliia Didenko <yu_lyan@ukr.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.108.0"
alembic = "^1.13.1"
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
uvicorn = "^0.25.0"
pydantic = {extras = ["email"], version = "^2.5.3"}
pydantic-extra-types = "^2.4.1"
phonenumbers = "^8.13.27"
libgravatar = "^1.0.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
fastapi-mail = "^1.4.1"
python-dotenv = "^1.0.1"
redis = "^5.0.1"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.38.0"
jinja2 = "^3.1.3"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"



[tool.poetry.group.test.dependencies]
pytest = "7.4.3"
pytest-asyncio = "^0.23.4"
httpx = "^0.26.0"
aiosqlite = "^0.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--doctest-modules"
testpaths = [
    "tests", ]
pythonpath = "."
filterwarnings = "ignore::DeprecationWarning"
